- внесены корректировки для работы PersonDetails

+++ core/src/main/java/com/test/application/core/domain/movieDetail/Rating.kt
@@ -5,5 +5,4 @@ data class Rating(
     val filmCritics: Double? = 0.0,
     val imdb: Double? = 0.0,
     val kp: Double? = 0.0,
-    val russianFilmCritics: Int? = 0
-)
+    val russianFilmCritics: Double? = 0.0)
