+++ remote_data/src/main/java/com/test/application/remote_data/dto/movieDetails/RatingDTO.kt
@@ -5,5 +5,5 @@ data class RatingDTO(
     val filmCritics: Double = 0.0,
     val imdb: Double = 0.0,
     val kp: Double = 0.0,
-    val russianFilmCritics: Int = 0
+    val russianFilmCritics: Double = 0.0
 )
